PYTORCH/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
Pillow>=9.0.0
PyYAML>=6.0
tqdm>=4.64.0
streamlit>=1.20.0
mlflow>=2.0.0
